+++ internal/store/gen/models.go
@@ -19,6 +19,7 @@ type Account struct {
 	Enabled         int64          `json:"enabled"`
 	CreatedAt       int64          `json:"created_at"`
 	UpdatedAt       int64          `json:"updated_at"`
+	AvatarUrl       string         `json:"avatar_url"`
 }
 
 type AccountCampaignLink struct {